+++ crates/libwebrtc-sys/src/cpp/bridge.cc
@@ -28,6 +28,11 @@
 
 namespace bridge {
 
+// Sets `libwebrtc` global log level.
+void set_webrtc_log_level(LoggingSeverity severity) {
+  webrtc::LogMessage::LogToDebug(severity);
+}
+
 // Indicates whether system audio capture is available on this platform.
 bool sys_audio_capture_is_available() { return SysAudioCaptureIsAvailable(); }