+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All user visible changes to this project will be documented in this file. This p
 ### Added
 
 - `setupAudioSessionManagement()` function for enabling/disabling automatic audio session management on [iOS]. ([#277])
+- `LogLevel` enum and `setLogLevel()` function to allow configuring native-side logging. ([#279])
 
 ### Fixed
 
@@ -26,6 +27,7 @@ All user visible changes to this project will be documented in this file. This p
 [#275]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/275
 [#277]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/277
 [#278]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/278
+[#279]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/279
 [16049300]: https://github.com/instrumentisto/medea-flutter-webrtc/commit/16049300183691fbbfafda26d39a263e6a89e1e0
 [143.0.7499.192]: https://github.com/instrumentisto/libwebrtc-bin/releases/tag/143.0.7499.192
 [openal-1.25.0]: https://github.com/kcat/openal-soft/releases/tag/1.25.0